db&query: add in affiliation for authors from ads
Request from Jess, as she wants to be able to log the EAO region of
the paper's first author. It might be better to log the affiliations
as a separate table, and then they could be identified into a master
list that is separately categorised by EAO region. This would be a lot
of work though -- scientists give affiliations pretty differently each
time.