comments: create comment class to hold a single comment
This refactored the db, and all the associated query/search/views
and templates. The paper view template now shows any previous comments
made on the paper (although not in a great format so far).

This got commited a little late, so I suspect some of the changes
for this may have been included in earlier commits...